預貯金を別ファイルへ切り出し
